Add participant route
