+++ src/api/participantsRouter.ts
@@ -1,4 +1,4 @@
-import express, { NextFunction, Request, Response } from 'express';
+import express, { Response } from 'express';
 import { z } from 'zod';
 
 import { Participant, ParticipantCreationPartial, ParticipantSchema } from './entities/Participant';
@@ -9,47 +9,19 @@ import { createNewUser, sendInviteEmail } from './services/kcUsersService';
 import {
   addSharingParticipants,
   deleteSharingParticipants,
+  enrichCurrentParticipant,
   getSharingParticipants,
+  hasParticipantAccess,
+  ParticipantRequest,
   sendNewParticipantEmail,
 } from './services/participantsService';
 import {
   insertSharingAuditTrails,
   updateAuditTrailsToProceed,
 } from './services/sharingAuditTrailService';
-import {
-  createUserInPortal,
-  findUserByEmail,
-  isUserBelongsToParticipant,
-} from './services/usersService';
+import { createUserInPortal, findUserByEmail } from './services/usersService';
 
 export const participantsRouter = express.Router();
-
-const idParser = z.object({
-  participantId: z.coerce.number(),
-});
-
-export interface ParticipantRequest extends Request {
-  participant?: Participant;
-}
-
-export const hasParticipantAccess = async (
-  req: ParticipantRequest,
-  res: Response,
-  next: NextFunction
-) => {
-  const { participantId } = idParser.parse(req.params);
-  const participant = await Participant.query().findById(participantId);
-  if (!participant) {
-    return res.status(404).send([{ message: 'The participant cannot be found.' }]);
-  }
-
-  if (!(await isUserBelongsToParticipant(req.auth?.payload?.email as string, participantId))) {
-    return res.status(403).send([{ message: 'You do not have permission to update participant.' }]);
-  }
-
-  req.participant = participant;
-  return next();
-};
 
 participantsRouter.get('/available', async (_req, res) => {
   const participants = await Participant.query().whereNotNull('siteId').withGraphFetched('types');
@@ -89,11 +61,17 @@ participantsRouter.post(
   hasParticipantAccess,
   async (req: ParticipantRequest, res: Response) => {
     try {
-      const { participantId } = idParser.parse(req.params);
+      const { participant } = req;
       const { firstName, lastName, email, role } = invitationParser.parse(req.body);
       const kcAdminClient = await getKcAdminClient();
       const user = await createNewUser(kcAdminClient, firstName, lastName, email);
-      await createUserInPortal({ email, role, participantId, firstName, lastName });
+      await createUserInPortal({
+        email,
+        role,
+        participantId: participant!.id,
+        firstName,
+        lastName,
+      });
       await sendInviteEmail(kcAdminClient, user);
       return res.sendStatus(201);
     } catch (err) {
@@ -108,6 +86,7 @@ participantsRouter.post(
 const participantParser = ParticipantSchema.pick({
   location: true,
 });
+
 participantsRouter.put(
   '/:participantId',
   hasParticipantAccess,
@@ -204,3 +183,6 @@ participantsRouter.post(
     return res.status(200).json(sharingParticipants);
   }
 );
+
+const currentParticipantRoute = express.Router();
+participantsRouter.use('/current', enrichCurrentParticipant, currentParticipantRoute);
